renamed to command dispatcher
