+++ Cargo.toml
@@ -163,7 +163,7 @@ pixi_allocator = { path = "crates/pixi_allocator" }
 pixi_build_frontend = { path = "crates/pixi_build_frontend" }
 pixi_build_type_conversions = { path = "crates/pixi_build_type_conversions" }
 pixi_build_types = { path = "crates/pixi_build_types" }
-pixi_command_queue = { path = "crates/pixi_command_queue" }
+pixi_command_dispatcher = { path = "crates/pixi_command_dispatcher" }
 pixi_config = { path = "crates/pixi_config" }
 pixi_consts = { path = "crates/pixi_consts" }
 pixi_default_versions = { path = "crates/pixi_default_versions" }
@@ -291,7 +291,7 @@ fs-err = { workspace = true, features = ["tokio"] }
 pixi_allocator = { workspace = true, optional = true }
 pixi_build_frontend = { workspace = true }
 pixi_build_types = { workspace = true }
-pixi_command_queue = { workspace = true }
+pixi_command_dispatcher = { workspace = true }
 pixi_config = { workspace = true }
 pixi_consts = { workspace = true }
 pixi_default_versions = { workspace = true }
