+++ crates/pixi_command_dispatcher/src/cache_dirs.rs
