+++ crates/pixi_command_dispatcher/src/event_reporter.rs
