--- crates/pixi_command_queue/Cargo.toml
@@ -1,48 +1,0 @@
-[package]
-authors.workspace = true
-edition.workspace = true
-homepage.workspace = true
-license.workspace = true
-name = "pixi_command_queue"
-readme.workspace = true
-repository.workspace = true
-version = "0.1.0"
-
-[dependencies]
-base64 = { workspace = true }
-chrono = { workspace = true }
-derive_more = { workspace = true, features = ["from"] }
-dirs = { workspace = true }
-futures = { workspace = true }
-itertools = { workspace = true }
-miette = { workspace = true }
-reqwest-middleware = { workspace = true }
-slotmap = { workspace = true }
-strsim = { workspace = true }
-thiserror = { workspace = true }
-tokio = { workspace = true, features = ["rt", "sync"] }
-tracing = { workspace = true }
-typed-path = { workspace = true }
-url = { workspace = true }
-xxhash-rust = { workspace = true, features = ["xxh3"] }
-serde = { workspace = true, features = ["derive"] }
-serde_json = { workspace = true }
-
-rattler_conda_types = { workspace = true }
-rattler_repodata_gateway = { workspace = true }
-rattler_solve = { workspace = true, features = ["resolvo"] }
-rattler_virtual_packages = { workspace = true }
-rattler = { workspace = true }
-rattler_shell = { workspace = true }
-
-pixi_build_frontend = { workspace = true }
-pixi_git = { workspace = true }
-pixi_glob = { workspace = true }
-pixi_record = { workspace = true }
-pixi_spec = { workspace = true }
-pixi_spec_containers = { workspace = true }
-pixi_utils = { workspace = true }
-
-[dev-dependencies]
-tokio = { workspace = true, features = ["macros"] }
-insta = { workspace = true, features = ["ron"] }
